refactor: move scraping utilities into test file for improved organization and maintainability
